+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Deprecate the `Security` class, use `Symfony\Bundle\SecurityBundle\Security\Security` instead
  * Change the signature of `TokenStorageInterface::setToken()` to `setToken(?TokenInterface $token)`
  * Deprecate calling `TokenStorage::setToken()` without arguments
+ * Add a `ChainUserChecker` to allow calling multiple user checkers for a firewall
 
 6.0
 ---